pyproject.toml
[tool.poetry]
name = "tapedrive"
version = "0.1.0"
description = "The selfhosted Podcast Archive"
authors = ["Jan Willhaus <mail@janwillhaus.de>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.7"
Django = "~2.2.10"
django-activity-stream = "^0.8"
django-background-tasks = "^1.2"
django-configurations = "^2.1"
djangorestframework = "^3.9"
djangorestframework_simplejwt = "^4.3"
dj-database-url = "^0.5.0"
bleach = "^3.1"
feedparser = "^6.0"
beautifulsoup4 = "^4.7"
Markdown = "^3.1"
Pillow = "^6.0"
requests = "^2.21"
python-dateutil = "^2.8"
html5lib = "^1.0"
whitenoise = "^4.1"
psycopg2 = "^2.9"
gunicorn = "*"
honcho = "*"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^1.11"
django-extensions = "^2.1"
pytest = "^4.4"
pytest-cov = "^2.7"
pytest-django = "^3.4"
pytest-mock = "^1.10"
pytest-vcr = "^1.0"
Werkzeug = "^0.15.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.267"
black = "^23.3.0"

[tool.ruff]
line-length = 120
target-version = "py39"
extend-select = [
    "I",   # isort
    "B",   # bugbear
    # "A",   # builtins FIXME: uncomment
    "ISC", # implicit-str-concat
    "Q",   # quotes
    "TID", # tidy-imports
    "C4",  # comprehensions
    "SIM", # simplify
    "C90", # mccabe
]
unfixable = [
    "B", # bugbear
]
ignore = [
    "SIM108", # if-else-block-instead-of-if-exp
]

[tool.ruff.mccabe]
max-complexity = 10

[tool.black]
line-length = 120
target-version = ['py37', 'py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  | frontend
  | build
  | dist
  | staticfiles
  | templates
  | assets
  | locale
  | node_modules
  | migrations
)/
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
